Implemented administration role and its management functionality

+++ Presentation/Controllers/TouristPlaceController.cs
@@ -1,10 +1,14 @@
+using System.Security.Claims;
 using Contracts;
+using Domain.Exceptions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services.Abstractions;
 
 namespace Presentation.Controllers;
 
 [ApiController]
+[Authorize]
 [Route("api/tourist-place")]
 public class TouristPlaceController : ControllerBase
 {
@@ -13,6 +17,7 @@ public class TouristPlaceController : ControllerBase
     public TouristPlaceController(IServiceManager serviceManager) => _serviceManager = serviceManager;
 
     [HttpGet]
+    [AllowAnonymous]
     public async Task<IActionResult> GetTouristPlaces(CancellationToken cancellationToken)
     {
         var touristPlaces = await _serviceManager.TouristPlaceService.GetAllAsync(cancellationToken);
@@ -21,6 +26,7 @@ public class TouristPlaceController : ControllerBase
     }
     
     [HttpGet("{searchParam}")]
+    [AllowAnonymous]
     public async Task<IActionResult> GetTouristPlacesBySearch(string searchParam, CancellationToken cancellationToken)
     {
         var touristPlaces = await _serviceManager.TouristPlaceService.GetBySearchAsync(searchParam, cancellationToken);
@@ -29,6 +35,7 @@ public class TouristPlaceController : ControllerBase
     }
 
     [HttpGet("{touristPlaceId:guid}")]
+    [AllowAnonymous]
     public async Task<IActionResult> GetTouristPlaceById(Guid touristPlaceId, CancellationToken cancellationToken)
     {
         var touristPlaceDto = await _serviceManager.TouristPlaceService.GetByIdAsync(touristPlaceId, cancellationToken);
@@ -39,6 +46,9 @@ public class TouristPlaceController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> CreateTouristPlace([FromBody] TouristPlaceForCreationDto touristPlaceForCreationDto)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         var touristPlaceDto = await _serviceManager.TouristPlaceService.CreateAsync(touristPlaceForCreationDto);
 
         return CreatedAtAction(nameof(GetTouristPlaceById), new { touristPlaceId = touristPlaceDto.Id }, touristPlaceDto);
@@ -47,6 +57,9 @@ public class TouristPlaceController : ControllerBase
     [HttpPut("{touristPlaceId:guid}")]
     public async Task<IActionResult> UpdateTouristPlace(Guid touristPlaceId, [FromBody] TouristPlaceForUpdateDto touristPlaceForUpdateDto, CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
         await _serviceManager.TouristPlaceService.UpdateAsync(touristPlaceId, touristPlaceForUpdateDto, cancellationToken);
 
         return NoContent();
@@ -55,6 +68,10 @@ public class TouristPlaceController : ControllerBase
     [HttpDelete("{touristPlaceId:guid}")]
     public async Task<IActionResult> DeleteTouristPlace(Guid touristPlaceId, CancellationToken cancellationToken)
     {
+        var userRole = User.FindFirst(ClaimTypes.Role)?.Value;
+        if (userRole != "Admin")
+            throw new InsufficientPrivilegeException("admin");
+        
         await _serviceManager.TouristPlaceService.DeleteAsync(touristPlaceId, cancellationToken);
 
         return NoContent();
